+++ src/main/java/repository/MongoDBRepository.java
@@ -5,7 +5,6 @@ import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.*;
-import domain.GameStatus;
 import domain.Position;
 import domain.Side;
 import domain.Turn;
